requirements.txt

streamlit==1.38.0
kokoro>=0.9.4
soundfile==0.12.1
numpy==1.24.3
torch>=2.0.0
python-dotenv==1.0.0